+++ testing/testing_zblas.cpp
@@ -64,7 +64,7 @@ int main( int argc, char** argv )
         n = opts.nsize[i];
         k = opts.ksize[i];
         printf("=========================================================================\n");
-        printf( "M %d, N %d, K %d\n", m, n, k );
+        printf( "M %d, N %d, K %d\n", (int) m, (int) n, (int) k );
         
         // allocate matrices
         // over-allocate so they can be any combination of {m,n,k} x {m,n,k}.
