+++ testing/testing_zungqr_m.cpp
@@ -123,11 +123,13 @@ int main( int argc, char** argv )
                 error = lapackf77_zlange("f", &m, &n, hR, &lda, work) / error;
                 
                 printf("%5d %5d %5d   %7.1f (%7.2f)   %7.1f (%7.2f)   %8.2e\n",
-                       m, n, k, cpu_perf, cpu_time, gpu_perf, gpu_time, error );
+                       (int) m, (int) n, (int) k,
+                       cpu_perf, cpu_time, gpu_perf, gpu_time, error );
             }
             else {
                 printf("%5d %5d %5d     ---   (  ---  )   %7.1f (%7.2f)     ---  \n",
-                       m, n, k, gpu_perf, gpu_time );
+                       (int) m, (int) n, (int) k,
+                       gpu_perf, gpu_time );
             }
             
             TESTING_HOSTFREE( hA     );
